separar simulador en primario y secundario
