+++ requirements.auxiliary.txt
@@ -1,2 +1,7 @@
+black
+ipython
+matplotlib
 numpy
+scipy
+sympy
 toml; python_version < '3.11'
